Extract buffer reading into custom file for reusability
